refactor(consumption): extract trip detail sections (Refs #563)
Trip detail screen previously held the summary card, share-payload
builder, body scaffold, and inner chart-section helper inline at 472
LOC. Splitting them into focused widgets brings the screen file under
the 300-LOC ceiling and lets each section be reasoned about (and
reused / tested) on its own.

- TripSummaryCard owns the 8-row summary + its formatter helpers
- TripDetailBody owns the scrollable layout + chart sections
- tripDetailSharePayload owns the JSON+CSV clipboard payload, with a
  visible-for-testing alias preserving the historical test entry point
